% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Function to fit user-specified GLM}
\usage{
fit_model(
  data,
  dbarts_x,
  dbarts_y,
  response_type = "continuous",
  response,
  main_effects = c(),
  interactions = c(),
  hierarchical = TRUE
)
}
\arguments{
\item{data}{A data frame of original data.}

\item{dbarts_x}{Matrix created using dbartsData - only covariates.}

\item{dbarts_y}{Matrix created using dbartsData - only outcome.}

\item{response_type}{Type of response - must be 'continuous' or 'binary'.}

\item{response}{Name of the response variable - must be character valued.}

\item{main_effects}{Vector of predictors to include as main effects - character valued.}

\item{interactions}{Vector of pairwise interactions to include - must have form 'x_i:x_j' and be character valued.}

\item{hierarchical}{TRUE/FALSE whether to force both main effects in model if corresponding pairwise interaction is included.}
}
\value{
A list of 2 objects. Object (1) is coefficient tibble, object (2) is the fit model object.
}
\description{
Function to fit user-specified GLM
}
\examples{
 library(MASS)
 library(dbarts)
 library(dplyr)
 library(janitor)
 data(birthwt)
 data = birthwt \%>\% dplyr::select(., -low)
 formula = bwt ~ .
 dbarts_data = dbartsData(formula, data)
 dbarts_x = clean_names(dbarts_data@x)
 dbarts_y = dbarts_data@y
 main_effects = c("smoke", "lwt", "race")
 interactions = c("race:smoke")
 model = fit_model(birthwt, dbarts_x, dbarts_y, response_type = "continuous",
                   response = "bwt", main_effects, interactions, hierarchical = TRUE)

}
