% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multiplier_thresh.R
\name{get_multiplier_thresh}
\alias{get_multiplier_thresh}
\title{Function to generate global SE threshold}
\usage{
get_multiplier_thresh(record, alpha = 0.01, use_means = TRUE)
}
\arguments{
\item{record}{Matrix of CIPs from get_posteriors()}

\item{alpha}{Determines 1 - alpha quantile within threshold}

\item{use_means}{TRUE/FALSE whether to use variable level CMD means. If FALSE means are set to 0.}
}
\value{
A data frame with 1 - alpha quantiles, means, sd's, and global SE threshold for each variable
}
\description{
Function to generate global SE threshold
}
\examples{
library(MASS)
library(dbarts)
library(dplyr)
library(janitor)
data(birthwt)
data = birthwt \%>\% dplyr::select(., -low)
formula = bwt ~ .
dbarts_data = dbartsData(formula, data)
dbarts_x = clean_names(dbarts_data@x)
dbarts_y = dbarts_data@y
vars = colnames(dbarts_x)
bart_run = run_bart(formula, data, vars = vars)

posteriors = get_posteriors(bart_run[[1]], vars, parallel::detectCores())
thresholds = get_multiplier_thresh(record = posteriors[[1]], alpha = 0.01, use_means = TRUE)

}
