% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thresholds.R
\name{get_thresholds}
\alias{get_thresholds}
\title{Function to get interaction and variable selection thresholds}
\usage{
get_thresholds(
  formula,
  data,
  num_trees = 10,
  num_samps = 5000,
  num_burn = 5000,
  num_thin = 5,
  prior_power = 4,
  prior_base = 0.99,
  num_chains = 4,
  num_threads_bart = min(num_chains, parallel::detectCores()),
  num_threads_wrangle = parallel::detectCores(),
  alpha_g = 0.01,
  alpha_l = 0.01,
  alpha_d = 0.01,
  alpha_g_vip = 0.01,
  num_run = 6,
  use_differences = TRUE,
  use_means = TRUE,
  diff_thresh,
  vars
)
}
\arguments{
\item{formula}{Object of class formula.}

\item{data}{A data frame of original data.}

\item{num_trees}{Number of BART trees per sample.}

\item{num_samps}{Number of posterior samples to take after burn-in.}

\item{num_burn}{Number of burn-in samples.}

\item{num_thin}{Keep every 'num_thin' draw.}

\item{prior_power}{Power parameter for tree prior.}

\item{prior_base}{Base parameter for tree prior.}

\item{num_chains}{Number of independent MCMC chains to run.}

\item{num_threads_bart}{Number of threads used to run BART - not recommended to be greater than 'num_chains'.}

\item{num_threads_wrangle}{Number of threads for post processing - can be larger than 'num_chains'.}

\item{alpha_g}{Alpha level for global threshold}

\item{alpha_l}{Alpha level for local thresholds}

\item{alpha_d}{Alpha level for global SE thresholds}

\item{alpha_g_vip}{Alpha level for variable selection threshold}

\item{num_run}{Number of null runs to do for threshold generations}

\item{use_differences}{TRUE/FALSE whether global SE threshold should be used for CMDs. If FALSE user must supply diff_thresh}

\item{use_means}{TRUE/FALSE whether to use means for CMD threshold. If false the variable level CMD means are set to 0.}

\item{diff_thresh}{A value for the difference threshold if use_differences set to FALSE.
use_differences set to TRUE overrides a value of diff_thresh}

\item{vars}{Column names resulting from dbartsData(data).}
}
\value{
A list of 4 or 5 objects depending on whether use_differences = TRUE. First object is a global CIP threshold.
Second object is a vector of local CIP thresholds. Third object is vector of CMD thresholds. Fourth object is a global
variable selection threshold. Fifth object (if use_differences = TRUE) is the return object from get_multiplier_thresh
}
\description{
Function to get interaction and variable selection thresholds
}
\examples{
library(MASS)
library(dbarts)
library(dplyr)
library(janitor)
data(birthwt)
data = birthwt \%>\% dplyr::select(., -low)
formula = bwt ~ .
dbarts_data = dbartsData(formula, data)
dbarts_x = clean_names(dbarts_data@x)
dbarts_y = dbarts_data@y
vars = colnames(dbarts_x)

thresholds = get_thresholds(formula, data, vars = vars)

}
\keyword{internal}
