% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vip.R
\name{get_vip}
\alias{get_vip}
\title{Function to get VIP from filtered tree data frame}
\usage{
get_vip(filtered_trees)
}
\arguments{
\item{filtered.trees}{}
}
\value{
A data frame of VIPs for each variable
}
\description{
Function to get VIP from filtered tree data frame
}
\examples{
library(MASS)
library(dbarts)
library(dplyr)
library(janitor)
data(birthwt)
data = birthwt \%>\% dplyr::select(., -low)
formula = bwt ~ .
dbarts_data = dbartsData(formula, data)
dbarts_x = clean_names(dbarts_data@x)
dbarts_y = dbarts_data@y
vars = colnames(dbarts_x)
bart_run = run_bart(formula, data, vars = vars)

vip = get_vip(filtered_trees = bart_run[[4]])

}
