% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not in operator}
\usage{
x %ni% vector
}
\arguments{
\item{x}{Values to be checked.}

\item{vector}{Values to be compared against.}
}
\value{
Logical vector indicating which elements of x are not in the vector
}
\description{
A convenience operator to negate \%in\%.
}
