% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not-in Operator (\verb{\%ni\%})}
\usage{
x \%ni\% vec
}
\arguments{
\item{x}{Vector or \code{NULL}; the values to be matched.}

\item{vec}{Vector or \code{NULL}; the values to be matched against.}
}
\value{
A logical vector of the same length as \code{x}, indicating which elements are \strong{not} in \code{vec}.
}
\description{
A convenience operator to negate \code{\%in\%}, returning a logical vector indicating if there is \strong{not} a match for each element of \code{x} in \code{vec}.
}
\examples{
v1 = 1:3
v2 = 1:10
v1 \%ni\% v2
v2 \%ni\% v1

}
