% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bart.R
\name{run_bart}
\alias{run_bart}
\title{Function to run dbarts::bart2}
\usage{
run_bart(
  formula,
  data,
  num_trees = 10,
  num_samps = 5000,
  num_burn = 5000,
  num_thin = 5,
  prior_power = 4,
  prior_base = 0.99,
  num_chains = 4,
  num_threads_bart = min(num_chains, parallel::detectCores()),
  num_threads_wrangle = parallel::detectCores(),
  vars
)
}
\arguments{
\item{formula}{Object of class formula.}

\item{data}{A data frame of original data.}

\item{num_trees}{Number of BART trees per sample.}

\item{num_samps}{Number of posterior samples to take after burn-in.}

\item{num_burn}{Number of burn-in samples.}

\item{num_thin}{Keep every 'num_thin' draw.}

\item{prior_power}{Power parameter for tree prior.}

\item{prior_base}{Base parameter for tree prior.}

\item{num_chains}{Number of independent MCMC chains to run.}

\item{num_threads_bart}{Number of threads used to run BART - not recommended to be greater than 'num_chains'.}

\item{num_threads_wrangle}{Number of threads for post processing - can be larger than 'num_chains'.}

\item{vars}{Column names resulting from dbartsData(data).}
}
\value{
A list of 2 objects. Object (1) is an indicator data frame of variable tree inclusion, object (2) is a data frame of posterior trees without terminal nodes.
}
\description{
Function to run dbarts::bart2
}
\examples{
library(MASS)
library(dbarts)
library(dplyr)
library(janitor)
data(birthwt)
data = birthwt \%>\% dplyr::select(., -low)
formula = bwt ~ .
dbarts_data = dbartsData(formula, data)
dbarts_x = clean_names(dbarts_data@x)
dbarts_y = dbarts_data@y
vars = colnames(dbarts_x)
bart_run = run_bart(formula, data, vars = vars)

}
