% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_thresholds.R
\name{test_thresholds}
\alias{test_thresholds}
\title{Function to test posterior CIPs, CMDs, MIPs against supplied thresholds}
\usage{
test_thresholds(
  global_thresh,
  local_thresholds,
  diff_thresholds,
  conditional_df,
  diff_df,
  marginals,
  method = "global"
)
}
\arguments{
\item{global_thresh}{A global CIP threshold value}

\item{local_thresholds}{A vector of local CIP thresholds - must be same length as marginals}

\item{diff_thresholds}{A vector of CMD thresholds - must be same length as marginals}

\item{conditional_df}{A data frame or matrix of posterior CIPs}

\item{diff_df}{A data frame or matrix of posterior CMDs}

\item{marginals}{A vector of posterior MIPs}

\item{method}{A value \%in\% c("global", "local", "both") to specify which selections are returned}
}
\value{
A list of selected interactions. If method \%in\% c("global", "local") the list will contain one data frame
of selected interactions and associated posterior info. If method = "both" a data frame will be returned for each method
}
\description{
Function to test posterior CIPs, CMDs, MIPs against supplied thresholds
}
\examples{
#' library(MASS)
library(dbarts)
library(dplyr)
library(janitor)
data(birthwt)
data = birthwt \%>\% dplyr::select(., -low)
formula = bwt ~ .
dbarts_data = dbartsData(formula, data)
dbarts_x = clean_names(dbarts_data@x)
dbarts_y = dbarts_data@y
vars = colnames(dbarts_x)

thresholds = get_thresholds(formula, data, vars = vars)
bart_run = real_bart_run(formula, data, vars = vars)

select = test_thresholds(thresholds[[1]], thresholds[[2]], thresholds[[3]],
                         bart_run[[1]], bart_run[[2]], bart_run[[3]],
                         method = "global")
                           
}
